+++ lib/sunmiprinterimports_web.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'dart:ui' as ui;
 import 'dart:typed_data';
 
 class SunmiPrinter {