+++ lib/sunmiscannerimports_web.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'dart:async';
 
 /// KeyAction enum for web implementation